pipeline/dags/domain_enrichment_dag.py
"""
PhishAware — Airflow DAG: Domain Enrichment Pipeline
DAG ID: domain_enrichment

This DAG runs every 30 minutes and:
1. Fetches domains from scan_logs that haven't been enriched yet
2. Looks up WHOIS data (domain age, registrar, country)
3. Queries VirusTotal API for link reputation
4. Checks DNS records (SPF, DMARC)
5. Computes a reputation score and writes to domain_reputation table
6. Updates scan_logs with enriched data
7. Logs the pipeline run to pipeline_runs table

This is the core DATA ENGINEERING pipeline of PhishAware.
"""

from datetime import datetime, timedelta
from airflow import DAG
from airflow.operators.python import PythonOperator
from airflow.utils.dates import days_ago
import logging

logger = logging.getLogger(__name__)

# ─── DAG Definition ──────────────────────────────────────────

default_args = {
    "owner": "phish-aware",
    "retries": 2,
    "retry_delay": timedelta(minutes=3),
    "email_on_failure": False,
}

dag = DAG(
    dag_id="domain_enrichment",
    default_args=default_args,
    description="Enrich scanned domains with WHOIS + VirusTotal + DNS data",
    schedule_interval="*/30 * * * *",  # every 30 minutes
    start_date=days_ago(1),
    catchup=False,
    tags=["phish-aware", "enrichment", "data-engineering"],
)

# ─── Task Functions ───────────────────────────────────────────

def extract_unenriched_domains(**context):
    """
    EXTRACT: Pull domains from scan_logs that need enrichment.
    Pushes domain list to XCom for downstream tasks.
    """
    # In production: query Supabase
    # from supabase import create_client
    # supabase = create_client(SUPABASE_URL, SUPABASE_KEY)
    # result = supabase.table("scan_logs")
    #   .select("sender_domain")
    #   .is_("enriched_at", "null")
    #   .limit(50)
    #   .execute()
    # domains = list(set([r["sender_domain"] for r in result.data]))
    
    domains = ["amaz0n-secure.xyz", "verify-bankofamerica.net"]  # demo
    logger.info(f"Extracted {len(domains)} unenriched domains")
    context["ti"].xcom_push(key="domains", value=domains)


def enrich_with_whois(**context):
    """
    TRANSFORM Step 1: Look up WHOIS data for each domain.
    """
    import whois
    domains = context["ti"].xcom_pull(key="domains", task_ids="extract_unenriched_domains")
    enriched = []
    
    for domain in domains:
        try:
            w = whois.whois(domain)
            creation = w.creation_date
            if isinstance(creation, list):
                creation = creation[0]
            age_days = (datetime.utcnow() - creation).days if creation else None
            
            enriched.append({
                "domain": domain,
                "age_days": age_days,
                "registrar": str(w.registrar)[:100] if w.registrar else None,
                "country": w.country,
            })
        except Exception as e:
            logger.warning(f"WHOIS failed for {domain}: {e}")
            enriched.append({"domain": domain, "age_days": None})
    
    context["ti"].xcom_push(key="whois_data", value=enriched)
    logger.info(f"WHOIS enrichment complete for {len(enriched)} domains")


def enrich_with_virustotal(**context):
    """
    TRANSFORM Step 2: Check domain reputation via VirusTotal API (free tier).
    """
    import requests, os
    VT_API_KEY = os.getenv("VIRUSTOTAL_API_KEY", "")
    domains = context["ti"].xcom_pull(key="domains", task_ids="extract_unenriched_domains")
    vt_data = {}
    
    for domain in domains:
        try:
            if not VT_API_KEY:
                vt_data[domain] = {"malicious": 0, "suspicious": 0, "clean": 75}
                continue
            
            resp = requests.get(
                f"https://www.virustotal.com/api/v3/domains/{domain}",
                headers={"x-apikey": VT_API_KEY},
                timeout=10
            )
            if resp.status_code == 200:
                stats = resp.json()["data"]["attributes"]["last_analysis_stats"]
                vt_data[domain] = {
                    "malicious": stats.get("malicious", 0),
                    "suspicious": stats.get("suspicious", 0),
                    "clean": stats.get("undetected", 0),
                }
        except Exception as e:
            logger.warning(f"VirusTotal failed for {domain}: {e}")
            vt_data[domain] = {"malicious": 0, "suspicious": 0, "clean": 0}
    
    context["ti"].xcom_push(key="vt_data", value=vt_data)


def check_dns_records(**context):
    """
    TRANSFORM Step 3: Check SPF and DMARC DNS records.
    """
    import dns.resolver
    domains = context["ti"].xcom_pull(key="domains", task_ids="extract_unenriched_domains")
    dns_data = {}
    
    for domain in domains:
        spf = False
        dmarc = False
        try:
            txt_records = dns.resolver.resolve(domain, "TXT")
            for record in txt_records:
                if "v=spf1" in str(record):
                    spf = True
            
            dmarc_records = dns.resolver.resolve(f"_dmarc.{domain}", "TXT")
            for record in dmarc_records:
                if "v=DMARC1" in str(record):
                    dmarc = True
        except Exception:
            pass
        
        dns_data[domain] = {"has_spf": spf, "has_dmarc": dmarc}
    
    context["ti"].xcom_push(key="dns_data", value=dns_data)


def compute_reputation_scores(**context):
    """
    TRANSFORM Step 4: Combine all enrichment data into a reputation score.
    """
    whois_data = context["ti"].xcom_pull(key="whois_data", task_ids="enrich_with_whois")
    vt_data = context["ti"].xcom_pull(key="vt_data", task_ids="enrich_with_virustotal")
    dns_data = context["ti"].xcom_pull(key="dns_data", task_ids="check_dns_records")
    
    scored = []
    for item in whois_data:
        domain = item["domain"]
        score = 100
        
        # Deduct for young domain
        age = item.get("age_days")
        if age is None:
            score -= 30
        elif age < 30:
            score -= 40
        elif age < 365:
            score -= 15
        
        # Deduct for VirusTotal hits
        vt = vt_data.get(domain, {})
        score -= vt.get("malicious", 0) * 5
        score -= vt.get("suspicious", 0) * 2
        
        # Deduct for missing DNS auth
        dns = dns_data.get(domain, {})
        if not dns.get("has_spf"):
            score -= 15
        if not dns.get("has_dmarc"):
            score -= 10
        
        scored.append({
            **item,
            **vt,
            **dns,
            "reputation_score": max(0, min(100, score))
        })
    
    context["ti"].xcom_push(key="scored_domains", value=scored)
    logger.info(f"Computed reputation scores for {len(scored)} domains")


def load_to_database(**context):
    """
    LOAD: Write enriched + scored domain data to Supabase.
    This is the final step of the ETL pipeline.
    """
    scored = context["ti"].xcom_pull(key="scored_domains", task_ids="compute_reputation_scores")
    
    # In production: upsert to Supabase
    # for record in scored:
    #     supabase.table("domain_reputation").upsert(record, on_conflict="domain").execute()
    
    logger.info(f"LOAD complete: {len(scored)} domain records written to Supabase")
    
    # Log pipeline run
    # supabase.table("pipeline_runs").insert({
    #     "dag_name": "domain_enrichment",
    #     "status": "success",
    #     "records_processed": len(scored),
    # }).execute()


# ─── Task Definitions ─────────────────────────────────────────

t1_extract = PythonOperator(
    task_id="extract_unenriched_domains",
    python_callable=extract_unenriched_domains,
    dag=dag,
)

t2_whois = PythonOperator(
    task_id="enrich_with_whois",
    python_callable=enrich_with_whois,
    dag=dag,
)

t3_vt = PythonOperator(
    task_id="enrich_with_virustotal",
    python_callable=enrich_with_virustotal,
    dag=dag,
)

t4_dns = PythonOperator(
    task_id="check_dns_records",
    python_callable=check_dns_records,
    dag=dag,
)

t5_score = PythonOperator(
    task_id="compute_reputation_scores",
    python_callable=compute_reputation_scores,
    dag=dag,
)

t6_load = PythonOperator(
    task_id="load_to_database",
    python_callable=load_to_database,
    dag=dag,
)

# ─── DAG Pipeline (ETL flow) ──────────────────────────────────
#
#  extract → whois ──┐
#                    ├──→ compute_scores → load
#           vt ──────┤
#           dns ─────┘
#

t1_extract >> [t2_whois, t3_vt, t4_dns] >> t5_score >> t6_load
