pipeline/requirements.txt
apache-airflow==2.9.0
apache-airflow-providers-http==4.10.0
supabase==2.4.0
python-whois==0.9.4
requests==2.31.0
dnspython==2.6.1
python-dotenv==1.0.1
